% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_eye_view.R
\name{birds_eye_view}
\alias{birds_eye_view}
\title{High level EDA: histograms, bar chart, and correlation heatmap}
\usage{
birds_eye_view(df, n = 20, var_list = NULL)
}
\arguments{
\item{df}{A data.frame. The dataframe on which visualizations will be created}

\item{n}{An integer. If passed, it will specify the number of bins for the histograms, defaults to 20}

\item{var_list}{A vector. If passed, it will limit visualizations to a subset of variables}
}
\value{
a list of the visualization objects generated created by the function
}
\description{
This function performs high level EDA on all, or a specific set of features displaying 3 different visualization sets: histograms for each numeric variable, a bar chart for each factor type categorical variable, and a correlation heatmap of the numeric variables.
}
\examples{
library(palmerpenguins)
results <- birds_eye_view(penguins)
}
