% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R, R/close_up.R
\name{clean_up}
\alias{clean_up}
\title{Takes a dataframe object and returns a cleaned version with rows containing any NaN values dropped. Inspects the clean dataframe and prints a list of potential outliers for each explanatory variable, based on the threshold distance of 3 standard deviations.}
\usage{
clean_up(df, N = 1)

clean_up(df, N = 1)
}
\arguments{
\item{df}{The dataframe on which the function will operate}

\item{N}{: The number of pairs of variables with strongest correlations to be displayed, defaults to 1}
}
\value{
the same dataframe with all the NaN's removed along with a list of potential outliers


}
\description{
Takes a dataframe object and returns a cleaned version with rows containing any NaN values dropped. Inspects the clean dataframe and prints a list of potential outliers for each explanatory variable, based on the threshold distance of 3 standard deviations.

Accepts a dataframe and the number of pairs of variables with strongest correlations. It creates scatterplots with a linear trend for each pair of variables with stronglest correlations.
}
\examples{
library(palmerpenguins)
results <- clean_up(penguins)
'**The following potenital outliers were detected:**
Variable X: 
[ 300, 301, 500, 1000 ]
Variable Y: 
[ 6.42, 6.44, 58.52, 60.22 ]'
library(palmerpenguins)
results <- clean_up(penguins)
}
