% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_up.R
\name{close_up}
\alias{close_up}
\title{Reveal linear relationships between two variables}
\usage{
close_up(df, n = 1)
}
\arguments{
\item{df}{The dataframe on which the function will operate}

\item{n}{The number of pairs of variables with strongest correlations to be displayed, defaults to 1}
}
\value{
chart
}
\description{
Accepts a dataframe and the number of pairs of variables with strongest correlations. It creates scatterplots with a linear trend for each pair of variables with stronglest correlations.
}
\examples{
library(palmerpenguins)
close_up(penguins)
}
