% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_df.R, R/shinyEDA.R
\docType{data}
\name{color_sets}
\alias{color_sets}
\title{Shiny EDA App}
\format{
An object of class \code{character} of length 29.

An object of class \code{character} of length 29.
}
\usage{
color_sets

color_sets
}
\arguments{
\item{data}{A data frame to feed into the Shiny app}
}
\value{



}
\description{
Shiny EDA App

Shiny EDA App
}
\examples{
#data(penguins, package = "modeldata")
#explore_df(penguins)

#data(penguins, package = "modeldata")
#explore_df(penguins)

}
\keyword{datasets}
