% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_df.R
\name{explore_df}
\alias{explore_df}
\title{Explore your data frame in an EDA Shiny App.}
\usage{
explore_df(data)
}
\arguments{
\item{data}{A data frame to feed into the Shiny app}
}
\value{
Shiny app with data loaded.
}
\description{
Explore your data frame in an EDA Shiny App.
}
\examples{
#data(penguins, package = "modeldata")
#explore_df(penguins)

}
