% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEDA.R
\name{shinyEDA}
\alias{shinyEDA}
\title{Shiny EDA App}
\usage{
shinyEDA(data)
}
\arguments{
\item{data}{A data frame to feed into the Shiny app}
}
\value{

}
\description{
Shiny EDA App
}
\examples{
#data(penguins, package = "modeldata")
#shinyEDA(penguins)

}
