% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{clean_up}
\alias{clean_up}
\title{Takes a dataframe object and returns a cleaned version with rows containing any NaN values dropped. Inspects the clean dataframe and prints a list of potential outliers for each explanatory variable.}
\usage{
clean_up(df)
}
\arguments{
\item{df}{The dataframe on which the function will operate}
}
\value{
the same dataframe with all the NaN's removed along with a list of potential outliers
}
\description{
Takes a dataframe object and returns a cleaned version with rows containing any NaN values dropped. Inspects the clean dataframe and prints a list of potential outliers for each explanatory variable.
}
\examples{
library(palmerpenguins)
results <- clean_up(penguins)
'**The following potential outliers are detected:**
$Variable X:
300, 301, 500, 1000
Variable Y:
6.42, 6.44, 58.52, 60.22'
}
