% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_df.R
\name{explore_df}
\alias{explore_df}
\title{Shiny EDA App}
\usage{
explore_df(data)
}
\arguments{
\item{data}{A data frame to feed into the Shiny app}

\item{output}{}
}
\value{



}
\description{
Shiny EDA App
}
\examples{
#data(penguins, package = "modeldata")
#explore_df(penguins)

}
