% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_suggestions.R
\name{summary_suggestions}
\alias{summary_suggestions}
\title{Takes a dataframe object and returns a list object comprising
of 3 dataframes and a list. The dataframes correspond to the
summary statistics of numeric and categorical variables each and
the proportion of unique values for categorical variables. The
nested list is of the categorical variables that exceed the threshold
for considering dropping variables with high unique values.}
\usage{
summary_suggestions(df, threshold)
}
\arguments{
\item{df}{The dataframe on which the function will operate}

\item{threshold}{A float. If passed, it will define the threshold for identifying the variables with high unique values, defaults to 0.8}
}
\value{

}
\description{
Takes a dataframe object and returns a list object comprising
of 3 dataframes and a list. The dataframes correspond to the
summary statistics of numeric and categorical variables each and
the proportion of unique values for categorical variables. The
nested list is of the categorical variables that exceed the threshold
for considering dropping variables with high unique values.
}
\examples{
library(palmerpenguins)
summary_suggestions(penguins)
}
